# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cahoots API
  description: API for Cahoots AI Development Team Platform
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.cahoots.ai
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    PasswordReset:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
        new_password:
          type: string
          format: password
          minLength: 8
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    ChangePasswordRequest:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          minLength: 8
...
paths:
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset email sent if account exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordReset'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Invalid reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password while logged in
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...